resolved conflicts

+++ proiect/src/main/java/com/example/demo/repositories/UserLimitRepository.java
@@ -2,6 +2,7 @@ package com.example.demo.repositories;
 
 
 import com.example.demo.domain.UserLimit;
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,5 +11,5 @@ import java.util.Optional;
 @Repository
 public interface UserLimitRepository extends JpaRepository<UserLimit, Integer> {
 
-    Optional<UserLimit> findByUserUserIdAndStatus(Long userId, String status);
+    Optional<UserLimit> findByUserUserIdAndStatus(Integer userId, String status);
 }
